requirements.txt
discord.py==1.3.3
praw==6.5.1